requirements.txt
absl-py==2.1.0
alembic==1.13.1
antspyx==0.4.2
apptools==5.2.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.3.3
category-encoders==2.5.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chart-studio==1.1.0
chemicals==1.1.5
click==8.1.7
cloudpickle==3.0.0
cmaes==0.10.0
colorama==0.4.6
colorcet==3.0.1
colorlog==6.8.2
comm==0.2.1
configobj==5.0.8
contourpy==1.2.0
convertdate==2.4.0
cycler==0.12.1
dask==2024.3.1
dask-expr==1.0.5
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
docrep==0.3.2
edt==2.4.0
einops==0.8.0
envisage==7.0.3
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
flatbuffers==24.3.25
fluids==1.0.25
fonttools==4.48.1
fsspec==2024.2.0
gast==0.5.4
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.1
h5py==3.10.0
hijri-converter==2.3.1
holidays==0.11.3.1
idna==3.6
imageio==2.34.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
intel-openmp==2023.2.4
ipydatawidgets==4.3.2
ipyevents==2.0.2
ipykernel==6.29.2
ipympl==0.9.3
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==7.8.1
itk-core==5.3.0
itk-filtering==5.3.0
itk-meshtopolydata==0.10.0
itk-numerics==5.3.0
itkwidgets==0.32.6
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab-widgets==1.1.7
jupyterlab_pygments==0.3.0
keras==2.14.0
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
lazy_loader==0.3
libclang==18.1.1
lightgbm==4.0.0
llvmlite==0.42.0
locket==1.0.0
lxml==5.1.0
Mako==1.3.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mayavi==4.8.1
mdurl==0.1.2
mistune==3.0.2
mkl==2023.2.0
ml-dtypes==0.2.0
mlxtend==0.23.1
monai-weekly==1.4.dev2409
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.2.0
nilearn==0.10.3
notebook==6.4.13
numba==0.59.0
numpy==1.26.4
oauthlib==3.2.2
opencv-python-headless==4.8.1.78
openpnm==3.4.1
openpyxl==3.1.2
opt-einsum==3.3.0
optuna==3.2.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
param==2.0.2
parso==0.8.3
partd==1.4.1
patsy==0.5.6
pillow==10.2.0
platformdirs==4.2.0
plotly==5.11.0
porespy==2.4.1
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyamg==5.0.1
pyarrow==15.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pyct==0.5.0
pyface==8.0.0
Pygments==2.17.2
PyMeeus==0.5.12
pypardiso==0.4.5
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
pytz==2024.1
PyWavelets==1.5.0
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
statsmodels==0.14.1
supervision==0.16.0
sympy==1.12
tbb==2021.11.0
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.0
thermo==0.2.27
threadpoolctl==3.2.0
tifffile==2024.2.12
tinycss2==1.2.1
toolz==0.12.1
torch==2.2.1+cu121
torchaudio==2.2.1+cu121
torchvision==0.17.1+cu121
tornado==6.4
tqdm==4.66.2
traitlets==5.6.0
traits==6.4.3
traitsui==8.0.0
traittypes==0.2.1
transforms3d==0.4.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
verstack==3.9.2
vtk==9.3.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
Werkzeug==3.0.1
widgetsnbextension==3.6.6
wrapt==1.14.1
xgboost==2.0.3
zipp==3.17.0
zstandard==0.22.0
